% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{api_first_unchanged}
\alias{api_first_unchanged}
\title{Return the first version where the function had the same API.}
\usage{
api_first_unchanged(fn, package, quick = TRUE, current_fn = NULL)
}
\arguments{
\item{fn}{Function name as a character string.}

\item{package}{Package. Alternatively you can specify the function name as e.g. \code{"package::function"}.}

\item{quick}{If \code{TRUE}, do a binary search that assumes API is never same, then different.}

\item{current_fn}{Current function for comparison. By default, \code{fn} in the current version of
the package. If provided, this must be an actual function, not a character string. You can use
\code{\link{get_fn_at}} to provide a suitable version.}
}
\value{
A version string.
}
\description{
Return the first version where the function had the same API.
}
\details{
A function's API is defined by its arguments as reported by \code{\link{formals}}.
}
\examples{
api_first_unchanged("to_latex", "huxtable")
}
