% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{api_same_at}
\alias{api_same_at}
\title{Test if a function's API is unchanged at a given version}
\usage{
api_same_at(fn, package, version = get_version_at_date(package, date),
  date = NULL, quiet = TRUE, current_fn = NULL, ...)
}
\arguments{
\item{fn}{Function name as a character string.}

\item{package}{Package. Alternatively, specify the function name as e.g. \code{"package::function"}.}

\item{version}{Version as a character string. If omitted, use the version available at \code{date}.}

\item{date}{Date, as a character string that can be read by \code{\link{as.Date}} e.g. "2016-01-01".}

\item{quiet}{Logical. Hide output from \code{install.packages}?}

\item{current_fn}{Current function for comparison. By default, \code{fn} in the current version of
the package (which is assumed to be available in a standard library location). If provided, this
must be an actual function, not a character string. You can use
\code{\link{get_fn_at}} for this.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{api_same_at} reports whether a function had the same API at a previous version or date.
}
\details{
If \code{fn} does not exist at \code{version}, \code{api_same_at} returns \code{FALSE} with a warning.
}
\examples{
\dontrun{
api_same_at("clipr::write_clip", version = "0.1.1")
# equivalently
api_same_at("write_clip", "clipr", "0.1.1")
}
}
