% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{api_unchanged_at}
\alias{api_unchanged_at}
\title{Is a function's API unchanged?}
\usage{
api_unchanged_at(fn, package, version = get_version_at_date(package, date),
  date = NULL, current_fn = NULL)
}
\arguments{
\item{fn}{Function name as a character string.}

\item{package}{Package. Alternatively you can specify the function name as e.g. \code{"package::function"}.}

\item{version}{Version as a character string. If omitted, use the version available at \code{date}.}

\item{date}{Date, as a character string that can be read by \code{\link{as.Date}} e.g. "2016-01-01".}

\item{current_fn}{Current function for comparison. By default, \code{fn} in the current version of
the package. If provided, this must be an actual function, not a character string. You can use
\code{\link{get_fn_at}} to provide a suitable version.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is a function's API unchanged?
}
\details{
If \code{fn} does not exist at \code{version}, this returns \code{FALSE} with a warning.
A function's API is defined by its arguments as reported by \code{\link{formals}}.
}
\examples{
\dontrun{
api_unchanged_at("huxreg", "huxtable", "2.0.0")
}
}
