% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementation.R
\name{apicheck-package}
\alias{apicheck-package}
\title{Basic details about the package}
\description{
Basic details about the package
}
\section{Warning}{

Do not try to use \code{apicheck} on itself. This will lead to fiery elephants in the sky.
}

