% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{available_versions}
\alias{available_versions}
\title{Report available versions}
\usage{
available_versions(package)
}
\arguments{
\item{package}{A single package name as a character string.}
}
\value{
A data frame with columns "version" and "date".
}
\description{
This is a simple wrapper round \code{\link[versions]{available.versions}}. It
returns packages ordered by date. Results are cached so as to
relieve pressure on the MRAN server. If \code{options("pastapi.use_cran") == FALSE},
then only versions available on MRAN (i.e. after 2014-09-17) will be returned;
otherwise older versions will be returned too.
}
\examples{
available_versions("clipr")

}
