% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementation.R
\name{cached_install}
\alias{cached_install}
\title{Install and/or load a version of a package}
\usage{
cached_install(package, version, return = c("namespace", "path"),
  cache = TRUE, quiet = TRUE, partial = TRUE, ...)
}
\arguments{
\item{package}{Package name.}

\item{version}{Version as a character string.}

\item{return}{Return the file "path" to the installed package, or the "namespace" object?}

\item{cache}{If \code{FALSE}, always reinstall the package.}

\item{quiet}{Logical. Try to minimize output from package installation. (Some output comes from \code{R CMD INSTALL} and may be unavoidable.)}

\item{partial}{Default \code{TRUE}. Passed to \code{\link[=loadNamespace]{loadNamespace()}}.}

\item{...}{Arguments passed to \code{\link[versions:install.versions]{versions::install.versions()}} or
\code{\link[remotes:install_version]{remotes::install_version()}}, and thence to \code{\link[=install.packages]{install.packages()}}.}
}
\value{
The namespace object or directory where the package is installed.
}
\description{
\code{cached_install} checks the package cache, installs the specified version if it is not
already installed, and loads the versioned package namespace.
}
\details{
If the package is already loaded, \code{cached_install} will first attempt
to unload it with a warning. This may not always work!

Note that the namespace is not attached. Partial loading is faster and safer when
you are (un)loading multiple versions, but does not export functions etc.
}
\examples{
\dontrun{
cached_install("clipr", "0.4.0")
}
}
