% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{clear_package_cache}
\alias{clear_package_cache}
\title{Delete all files from the package cache}
\usage{
clear_package_cache()
}
\value{
TRUE if all files and directories could be removed, FALSE otherwise.
}
\description{
This attempts to delete \emph{everything} in the package cache.
}
\details{
The package cache is under the directory specified by \code{getOption("apicheck.lib_dir")},
or, if that is unset, in a per-session temporary directory. You should use \code{\link{set_lib_dir}}
to change this rather than setting the option directly. Within this directory, subdirectories
are named like \code{package-version}, e.g. \code{longurl-0.3.0}. Within these subdirectories are
the actual installed libraries.
}
\examples{
\dontrun{
clear_package_cache()
}
}
\seealso{
Other utility functions: \code{\link{set_lib_dir}}
}
