% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{compare_versions}
\alias{compare_versions}
\title{Compare versions of a package and report changed functions and APIs}
\usage{
compare_versions(package, version = previous_version(package),
  version2 = NULL, quiet = TRUE, ...)
}
\arguments{
\item{package}{Package name.}

\item{version}{First version to compare. If \code{NULL}, use the previous available version.}

\item{version2}{Second version to compare. If \code{NULL}, use the current version as installed.}

\item{quiet}{Logical. Try to minimize output from package installation. (Some output comes from \code{R CMD INSTALL} and may be unavoidable.)}

\item{...}{Arguments passed to \code{\link[versions:install.versions]{versions::install.versions()}} or
\code{\link[remotes:install_version]{remotes::install_version()}}, and thence to \code{\link[=install.packages]{install.packages()}}.}
}
\value{
A data frame reporting functions that have been "Added", "Removed" or had "API changed",
and details of function arguments.
}
\description{
Compare versions of a package and report changed functions and APIs
}
\examples{
\dontrun{
compare_versions("clipr", "0.2.1", "0.3.0")
}
}
