% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{fn_exists_at}
\alias{fn_exists_at}
\title{Test if a function exists at a given version}
\usage{
fn_exists_at(fn, package, version = get_version_at_date(package, date),
  date = NULL)
}
\arguments{
\item{fn}{Function name as a character string.}

\item{package}{Package. Alternatively, specify the function name as e.g. \code{"package::function"}.}

\item{version}{Version as a character string. If omitted, use the version available at \code{date}.}

\item{date}{Date, as a character string that can be read by \code{\link{as.Date}} e.g. "2016-01-01".}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{fn_exists_at} reports whether a function exists at a specific previous version or date.
}
\examples{
\dontrun{
fn_exists_at("clipr::dr_clipr", version = "0.3.1")
# or
fn_exists_at("dr_clipr", "clipr", "0.3.1")
}
}
