% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R, R/search.R
\name{fn_exists_at}
\alias{fn_exists_at}
\alias{fn_first_exists}
\title{Compare function existence across package versions}
\usage{
fn_exists_at(fn, package, version = get_version_at_date(package, date),
  date = NULL)

fn_first_exists(fn, package, search = c("binary", "forward", "backward",
  "all"), report = c("full", "brief"))
}
\arguments{
\item{fn}{Function name as a character string.}

\item{package}{Package. Alternatively, specify the function name as e.g. \code{"package::function"}.}

\item{version}{Version as a character string. If omitted, use the version available at \code{date}.}

\item{date}{Date, as a character string that can be read by \code{\link{as.Date}} e.g. "2016-01-01".}

\item{search}{"binary", "forward", "backward" or "all". See Details.}

\item{report}{"brief" or "full". See Value.}
}
\value{
\code{fn_exists_at} returns \code{TRUE} or \code{FALSE}.
}
\description{
\code{fn_first_exists} reports the first package version where a function exists. \code{fn_exists_at} reports
whether a function exists at a specific previous version.
}
\details{
This function downloads and installs multiple versions from MRAN, so it is likely to be slow.
}
\examples{
\dontrun{
fn_exists_at("read.arff", "foreign", version = "0.8-19")
}
\dontrun{
fn_first_exists('read.dta', 'foreign')
}
}
