% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{fun_at}
\alias{fun_at}
\title{Retrieve a function from a package version}
\usage{
fun_at(fun, version = version_at_date(package, date), package, date = NULL,
  quiet = TRUE, allow_core = FALSE, ...)
}
\arguments{
\item{fun}{Function name as a character string. \code{fun} can be an S3 method; S4 methods aren't yet supported.}

\item{version}{Version as a character string. If omitted, use the version available at \code{date}.}

\item{package}{Package. Alternatively, specify the function name as e.g. \code{"package::function"}.}

\item{date}{Date, as a character string that can be read by \code{\link[=as.Date]{as.Date()}} e.g. "2016-01-01".}

\item{quiet}{Logical. Try to minimize output from package installation. (Some output comes
from \code{R CMD INSTALL} and may be unavoidable.)}

\item{allow_core}{See below.}

\item{...}{Arguments passed to \code{\link[versions:install.versions]{versions::install.versions()}} or
\code{\link[remotes:install_version]{remotes::install_version()}}, and thence to \code{\link[=install.packages]{install.packages()}}. \code{Ncpus} may be useful.}
}
\value{
The function itself.
}
\description{
Retrieve a function from a package version
}
\details{
By default, you cannot get functions from previous versions of R core packages, which are not available
on CRAN/MRAN. If \code{allow_core} is \code{TRUE}, then a function will be returned from the
\href{https://github.com/hughjonesd/rcheology}{rcheology} dataset, with a null body but the same formal
arguments as the historical function.
}
\examples{
\dontrun{
fun_at("write_clip", "clipr", "0.1.1")
}
}
