% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{fun_exists_at}
\alias{fun_exists_at}
\title{Test if a function exists at a given version}
\usage{
fun_exists_at(fun, version = version_at_date(package, date), package,
  date = NULL, quiet = TRUE, ...)
}
\arguments{
\item{fun}{Function name as a character string. \code{fun} can be an S3 method; S4 methods aren't yet supported.}

\item{version}{Version as a character string. If omitted, use the version available at \code{date}.}

\item{package}{Package. Alternatively, specify the function name as e.g. \code{"package::function"}.}

\item{date}{Date, as a character string that can be read by \code{\link[=as.Date]{as.Date()}} e.g. "2016-01-01".}

\item{quiet}{Logical. Try to minimize output from package installation. (Some output comes from \code{R CMD INSTALL} and may be unavoidable.)}

\item{...}{Arguments passed to \code{\link[versions:install.versions]{versions::install.versions()}} or
\code{\link[remotes:install_version]{remotes::install_version()}}, and thence to \code{\link[=install.packages]{install.packages()}}.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{fun_exists_at} reports whether a function exists (i.e. is exported) from a package at a specific previous version
or date.
}
\examples{
\dontrun{
fun_exists_at("clipr::dr_clipr", version = "0.3.1")
# or
fun_exists_at("dr_clipr", "clipr", "0.3.1")
}
}
