% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementation.R
\name{get_help_at}
\alias{get_help_at}
\title{Get text help for a function at a package version}
\usage{
get_help_at(fn, version = get_version_at_date(package, date), package,
  date = NULL, quiet = TRUE, ...)
}
\arguments{
\item{fn}{Function name as a character string.}

\item{version}{Version as a character string. If omitted, use the version available at \code{date}.}

\item{package}{Package. Alternatively, specify the function name as e.g. \code{"package::function"}.}

\item{date}{Date, as a character string that can be read by \code{\link{as.Date}} e.g. "2016-01-01".}

\item{quiet}{Logical. Try to minimize output from package installation. (Some output comes from \code{R CMD INSTALL} and may be unavoidable.)}

\item{...}{Arguments passed to \code{\link[versions]{install.versions}} or
\code{\link[remotes]{install_version}}, and thence to \code{\link{install.packages}}.}
}
\value{
The help object in text format.
}
\description{
Get text help for a function at a package version
}
\examples{
\dontrun{
get_help_at("clipr::write_clip", "0.1.1")
}
}
\seealso{
\code{\link{help}}
}
