% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{get_version_at_date}
\alias{get_version_at_date}
\title{Return the current version of a package at a given date}
\usage{
get_version_at_date(package, date)
}
\arguments{
\item{package}{Package. Alternatively, specify the function name as e.g. \code{"package::function"}.}

\item{date}{Date, as a character string that can be read by \code{\link{as.Date}} e.g. "2016-01-01".}
}
\value{
A version string.
}
\description{
Return the current version of a package at a given date
}
\examples{
\dontrun{
get_version_at_date("huxtable", "2017-01-01")
}
}
