% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementation.R
\name{load_version_namespace}
\alias{load_version_namespace}
\title{Load the namespace from a version of a package}
\usage{
load_version_namespace(package, version, cache = TRUE, quiet = TRUE, ...)
}
\arguments{
\item{package}{Package name.}

\item{version}{Version as a character string.}

\item{cache}{If \code{FALSE}, always try to reinstall the package.}

\item{quiet}{Logical. Try to minimize output from package installation. (Some output comes from \code{R CMD INSTALL} and may be unavoidable.)}

\item{...}{Arguments passed to \code{\link[versions]{install.versions}} or
\code{\link[remotes]{install_version}}, and thence to \code{\link{install.packages}}.}
}
\value{
The namespace object.
}
\description{
Load the namespace from a version of a package
}
\details{
If the package is not found in the package cache, it will be downloaded and
installed there.

If the package is already loaded, \code{load_version_namespace} will first attempt
to unload it with a warning. This may not always work!

Note that the namespace is not attached.
}
\examples{
\dontrun{
load_version_namespace("clipr", "0.4.0")
}
}
