% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{mran_versions}
\alias{mran_versions}
\title{Report versions available on MRAN}
\usage{
mran_versions(package)
}
\arguments{
\item{package}{A single package name as a character string.}
}
\value{
A data frame with columns "version", "date" and "available" (always \code{TRUE}).
}
\description{
This is a simple wrapper round \code{\link[versions]{available.versions}}. It
returns only packages on MRAN and is ordered by date. Results are cached so as to
relieve pressure on the MRAN server.
}
\examples{
mran_versions("clipr")

}
