% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{pastapi-package}
\alias{pastapi-package}
\title{Basic details about the package}
\description{
This is a small package to check when functions were introduced in packages and/or APIs changed.
It uses the \code{versions} package to install different versions of a package
from \href{https://mran.microsoft.com/}{MRAN}.
}
\details{
Packages are cached within a session. To cache packages across sessions, use
\code{\link{set_lib_dir}} to point to a persistent directory.

Be aware that functions can take a long time to return, as different versions of a package are
installed and/or loaded.

Also, be aware that namespace loading and unloading can be unreliable. If this happens to you, try
restarting your session.
}
