% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{set_lib_dir}
\alias{set_lib_dir}
\alias{get_lib_dir}
\title{Specify library location}
\usage{
set_lib_dir(lib_dir)

get_lib_dir()
}
\arguments{
\item{lib_dir}{Path to a directory, or \code{NULL} to unset.}
}
\value{
The old library location, invisibly. By default this is a subdirectory of \code{\link{tempdir}}.

\code{get_lib_dir} returns the library location.
}
\description{
This specifies where libraries will be downloaded to,
and resets the cache of installed library locations.
}
\details{
If \code{lib_dir} does not exist it will be created. If you set it to \code{NULL}, a subdirectory
of \code{tempdir()} will be used.
}
\examples{
\dontrun{
set_lib_dir("~/.pastapi")
}
get_lib_dir()
}
\seealso{
Other utility functions: \code{\link{clear_package_cache}}
}
