% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{set_pastapi_lib_dir}
\alias{set_pastapi_lib_dir}
\title{Specify library location}
\usage{
set_pastapi_lib_dir(lib_dir)
}
\arguments{
\item{lib_dir}{Path to a directory.}
}
\value{
The old library location. By default this is a subdirectory of \code{\link{tempdir()}}.
}
\description{
This specifies where libraries will be downloaded to,
and resets the cache of installed library locations.
}
\details{
If \code{lib_dir} does not exist it will be created.
}
\examples{
\dontrun{
set_pastapi_lib_dir("~/.pastapi")
}
}
