% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{apicheck-package}
\alias{apicheck-package}
\title{apicheck: check function APIs in different versions of packages}
\description{
This is a small package to check when functions were introduced and/or APIs changed in packages.
It automatically installs different versions of a package in a separate directory and loads them
without attaching them.
}
\details{
Packages are cached within a session. To cache packages across sessions, use
\code{\link[=set_lib_dir]{set_lib_dir()}} to point to a persistent directory.

By default, \code{apicheck`` uses the}remotes\code{package to install source versions from CRAN. Alternatively, it can use the}versions`` package to install different versions of a package
from \href{MRAN}{https://mran.microsoft.com/}. To do this set \code{options(apicheck.use_mran = TRUE)}.

Be aware that functions can take a long time to return, as different versions of a package are
installed and/or loaded.

Also, be aware that namespace loading and unloading can be unreliable. If this happens to you, try
restarting your session.
}
\section{Warning}{

Do not try to use \code{apicheck} on itself. This will lead to fiery elephants in the sky.
}

