% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{available_versions}
\alias{available_versions}
\title{Report available versions}
\usage{
available_versions(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
A data frame with columns "version" and "date".
}
\description{
This returns packages ordered by date, using either \href{https://mran.microsoft.com}{MRAN} or
\href{http://crandb.r-pkg.org}{metacran} (or \href{https::/github.com/hughjonesd/rcheology}{rcheology}
for R core packages).
Results are cached so as to
relieve pressure on the server. If \code{options("apicheck.use_mran")} is \code{TRUE} (and \code{package} is non-core),
then only versions available on MRAN (i.e. after 2014-09-17) will be returned;
otherwise older versions will be returned too.
}
\section{Speed}{

In my limited experience, metacran is much faster. YMMV.
}

\examples{
\dontrun{
available_versions("clipr")
}

}
