% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementation.R
\name{cached_install}
\alias{cached_install}
\title{Return a package version's location or namespace, possibly installing it}
\usage{
cached_install(package, version, return = c("namespace", "path"),
  cache = TRUE, quiet = TRUE, partial = TRUE, ...)
}
\arguments{
\item{package}{Package name.}

\item{version}{Version as a character string.}

\item{return}{Return the file "path" to the installed package, or the "namespace" object?}

\item{cache}{If \code{FALSE}, always try to reinstall the package.}

\item{quiet}{Logical. Try to minimize output from package installation. (Some output comes from \code{R CMD INSTALL} and may be unavoidable.)}

\item{partial}{Default \code{TRUE}. Passed to \code{\link{loadNamespace}}.}

\item{...}{Arguments passed to \code{\link[versions]{install.versions}} or
\code{\link[remotes]{install_version}}, and thence to \code{\link{install.packages}}.}
}
\value{
The namespace object or directory where the package is installed.
}
\description{
Return a package version's location or namespace, possibly installing it
}
\details{
If the package is not found in the package cache, it will be downloaded and
installed there.

If the package is already loaded, \code{cached_install} will first attempt
to unload it with a warning. This may not always work!

Note that the namespace is not attached. Partial loading is faster and safer when
you are (un)loading multiple versions, but does not export functions etc.
}
\examples{
\dontrun{
cached_install("clipr", "0.4.0")
}
}
