% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{call_with_namespace}
\alias{call_with_namespace}
\title{Loads a package namespace at a particular version and runs an arbitrary function}
\usage{
call_with_namespace(package, version, test, quiet = TRUE, ...)
}
\arguments{
\item{package}{Package name.}

\item{version}{Version as a character string.}

\item{test}{A one-argument function. See Details.}

\item{quiet}{Logical. Try to minimize output from package installation. (Some output comes from \code{R CMD INSTALL} and may be unavoidable.)}

\item{...}{Arguments passed to \code{\link[versions]{install.versions}} or
\code{\link[remotes]{install_version}}, and thence to \code{\link{install.packages}}.}
}
\value{
The value returned by \code{test}.
}
\description{
Loads a package namespace at a particular version and runs an arbitrary function
}
\details{
The package is downloaded and installed if necessary, and its namespace is loaded. Then the
\code{test(ns)} is called with the namespace object, and its value is returned. On exit, the
namespace is unloaded, hopefully leaving your environment clean.
}
\examples{
\dontrun{
can_it_expand_urls <- function (namespace) "expand_urls" \%in\% names(namespace)
call_with_namespace("longurl", "0.3.0", test = can_it_expand_urls)
}
}
