% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{clear_package_cache}
\alias{clear_package_cache}
\title{Delete all files from the package cache}
\usage{
clear_package_cache()
}
\value{
TRUE if all files and directories could be removed, FALSE otherwise.
}
\description{
This attempts to delete \emph{everything} in the package cache, which is stored
either in \code{getOption("pastapi.lib_dir")} or, if that is unset, in a
per-session temporary directory.
}
\examples{
\dontrun{
clear_package_cache()
}
}
