% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{compare_versions}
\alias{compare_versions}
\alias{print.versions_report}
\alias{summary.versions_report}
\title{Compare versions of a package}
\usage{
compare_versions(package, version = previous_version(package, version2),
  version2, methods = FALSE, quiet = TRUE, ...)

\method{print}{versions_report}(x, ...)

\method{summary}{versions_report}(object, ...)
}
\arguments{
\item{package}{Package name.}

\item{version}{First version to compare. By default, the previous version to that currently
installed.}

\item{version2}{Second version to compare. By default, the current installed version.}

\item{methods}{Logical: include non-exported S3 methods?}

\item{quiet}{Logical. Try to minimize output from package installation. (Some output comes
from \code{R CMD INSTALL} and may be unavoidable.)}

\item{...}{Arguments passed to \code{\link[versions:install.versions]{versions::install.versions()}} or
\code{\link[remotes:install_version]{remotes::install_version()}}, and thence to \code{\link[=install.packages]{install.packages()}}. \code{Ncpus} may be useful.}

\item{x, object}{An object of class \code{versions_report}.}
}
\value{
\code{compare_versions} returns a data frame of class \code{versions_report},
reporting functions that have been "Added", "Removed" or had "API changed",
and details of function arguments. Extra information is in the \code{"package"} and \code{"versions"}
attributes.

\code{summary} returns a string representation of changed function arguments.
}
\description{
\code{compare_versions} reports how functions and APIs changed between versions of a package.
}
\examples{
\dontrun{
compare_versions("clipr", "0.2.1", "0.3.0")
}
}
