% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{fn_exists_at}
\alias{fn_exists_at}
\title{Does a function exist at a given version or date?}
\usage{
fn_exists_at(fn, package, version = get_version_at_date(package, date),
  date = NULL)
}
\arguments{
\item{fn}{Function name as a character string.}

\item{package}{Package. Alternatively you can specify the function name as e.g. \code{"package::function"}.}

\item{version}{Version as a character string. If omitted, use the version available at \code{date}.}

\item{date}{Date, as a character string that can be read by \code{\link{as.Date}} e.g. "2016-01-01".}
}
\value{
TRUE or FALSE.
}
\description{
Does a function exist at a given version or date?
}
\examples{
\\notrun{
fn_exists_at('as_Workbook', 'huxtable', version = '3.0.0')
}
}
