% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apihistory.R
\name{fn_first_exists}
\alias{fn_first_exists}
\title{Find the first version when a function existed}
\usage{
fn_first_exists(fn, package, quick = TRUE)
}
\arguments{
\item{fn}{Function name.}

\item{package}{Package name.}

\item{quick}{If TRUE, perform a binary search. This assumes that once added, a
function does not go away.}
}
\value{
A version string.
}
\description{
Find the first version when a function existed
}
\examples{
fn_first_exists('as_Workbook', 'huxtable')
}
