% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{fun_at}
\alias{fun_at}
\title{Retrieve a function from a particular package version}
\usage{
fun_at(fun, version = version_at_date(package, date), package, date = NULL,
  quiet = TRUE, ...)
}
\arguments{
\item{fun}{Function name as a character string. \code{fun} can be an S3 method; S4 methods aren't yet supported.}

\item{version}{Version as a character string. If omitted, use the version available at \code{date}.}

\item{package}{Package. Alternatively, specify the function name as e.g. \code{"package::function"}.}

\item{date}{Date, as a character string that can be read by \code{\link[=as.Date]{as.Date()}} e.g. "2016-01-01".}

\item{quiet}{Logical. Try to minimize output from package installation. (Some output comes from \code{R CMD INSTALL} and may be unavoidable.)}

\item{...}{Arguments passed to \code{\link[versions:install.versions]{versions::install.versions()}} or
\code{\link[remotes:install_version]{remotes::install_version()}}, and thence to \code{\link[=install.packages]{install.packages()}}.}
}
\value{
The function itself.
}
\description{
Retrieve a function from a particular package version
}
\examples{
\dontrun{
fun_at("write_clip", "clipr", "0.1.1")
}
}
