% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{get_fn_at}
\alias{get_fn_at}
\title{Retrieve a function from a particular package version}
\usage{
get_fn_at(fn, package, version = get_version_at_date(package, date),
  date = NULL)
}
\arguments{
\item{fn}{Function name as a character string.}

\item{package}{Package. Alternatively, specify the function name as e.g. \code{"package::function"}.}

\item{version}{Version as a character string. If omitted, use the version available at \code{date}.}

\item{date}{Date, as a character string that can be read by \code{\link{as.Date}} e.g. "2016-01-01".}
}
\value{
The function itself.
}
\description{
Retrieve a function from a particular package version
}
\examples{
\dontrun{
get_fn_at("huxreg", "huxtable", "2.0.0")
}
}
