% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{get_version_at_date}
\alias{get_version_at_date}
\title{Return the current version of a package at a given date}
\usage{
get_version_at_date(package, date)
}
\arguments{
\item{package}{A package as a string.}

\item{date}{A date.}
}
\value{
A version string.
}
\description{
Return the current version of a package at a given date
}
\examples{
\dontrun{
get_version_at_date("huxtable", "2017-01-01")
}
}
