% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastapi.R
\name{load_version_namespace}
\alias{load_version_namespace}
\title{Load the namespace from a version of a package}
\usage{
load_version_namespace(package, version, cache = TRUE)
}
\arguments{
\item{package}{Package name.}

\item{version}{Version as a character string.}

\item{cache}{If FALSE, always try to reinstall the package.}
}
\value{
The namespace object.
}
\description{
Load the namespace from a version of a package
}
\details{
If the package is not found in the package cache, it will be downloaded and
installed there.
}
\examples{
\dontrun{
load_version_namespace("clipr", "0.4.0")
}
}
