% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{set_lib_dir}
\alias{set_lib_dir}
\alias{get_lib_dir}
\title{Specify library location}
\usage{
set_lib_dir(lib_dir, create = FALSE)

get_lib_dir()
}
\arguments{
\item{lib_dir}{Path to a directory, or \code{NULL} to unset.}

\item{create}{Logical. Try to create the directory if it doesn't exist.}
}
\value{
\code{set_lib_dir} invisibly returns the old library location, or \code{NULL} if none was set in options.

\code{get_lib_dir} returns the actual library location, whether or not an option has been set.
}
\description{
This specifies where libraries will be downloaded to,
and resets the cache of installed library locations.
}
\details{
If \code{lib_dir} is set to \code{NULL}, a subdirectory of \code{tempdir()} will be used.
\code{lib_dir} will be normalized via \code{\link{normalizePath}}.
}
\examples{
\dontrun{
set_lib_dir("~/.apicheck")
}
get_lib_dir()
}
\seealso{
Other utility functions: \code{\link{clear_package_cache}}
}
